App 投稿者専用機能化

+++ resources/views/posts/show.blade.php
@@ -30,12 +30,15 @@
                 </div>
             @endforeach
         </div>
-        <div class="edit"><a href="/posts/{{ $post->id }}/edit">edit</a></div>
-        <form action="/posts/{{ $post->id }}" id="form_{{ $post->id }}" method="post">
-            @csrf
-            @method('DELETE')
-            <button type="button" onclick="deletePost({{ $post->id }})">delete</button> 
-        </form>
+        <!--投稿者専用機能-->
+        @if($post->user->id===$Auth)
+            <div class="edit"><a href="/posts/{{ $post->id }}/edit">edit</a></div>
+            <form action="/posts/{{ $post->id }}" id="form_{{ $post->id }}" method="post">
+                @csrf
+                @method('DELETE')
+                <button type="button" onclick="deletePost({{ $post->id }})">delete</button> 
+            </form>
+        @endif
         <br>
         <hr>
         <div class="comment">
@@ -43,11 +46,13 @@
                 <a href="/users/{{$post->user->id}}">{{$comment->user->name}}より</a>
                 <small>{{$comment->created_at}}</small>
                 <p>{{$comment->body}}</p>
-                <form action="/posts/comments/{{$comment->id}}" id="form_{{ $comment->id }}" method="post">
-                    @csrf
-                    @method('DELETE')
-                    <button type="button" onclick="deleteComment({{ $comment->id }})">delete</button> 
-                </form>
+                @if($comment->user->id===$Auth)
+                    <form action="/posts/comments/{{$comment->id}}" id="form_{{ $comment->id }}" method="post">
+                        @csrf
+                        @method('DELETE')
+                        <button type="button" onclick="deleteComment({{ $comment->id }})">delete</button> 
+                    </form>
+                @endif
                 <hr>
             @endforeach
